Rename: 긴 클래스명 단축
'장기 미수령 택배 관리대장 게시글' 을 의미하는
UnDeliveredParcelRegisterPost 를 UnDeliveredParcelPost 로 변경
