+++ src/main/java/com/dominest/dominestbackend/api/post/undeliveredparcel/controller/UnDeliveredParcelController.java
@@ -1,7 +1,8 @@
 package com.dominest.dominestbackend.api.post.undeliveredparcel.controller;
 
 import com.dominest.dominestbackend.api.common.ResTemplate;
-import com.dominest.dominestbackend.domain.post.undeliveredparcelregister.UnDeliveredParcelRegisterPostService;
+import com.dominest.dominestbackend.domain.post.undeliveredparcel.UndeliveredParcelPostService;
+import com.dominest.dominestbackend.domain.post.undeliveredparcel.component.UndeliveredParcelService;
 import com.dominest.dominestbackend.global.util.PrincipalUtil;
 import lombok.RequiredArgsConstructor;
 import org.springframework.http.HttpStatus;
@@ -16,7 +17,8 @@ import java.security.Principal;
 @RequiredArgsConstructor
 @RestController
 public class UnDeliveredParcelController {
-    private final UnDeliveredParcelRegisterPostService unDeliParcelService;
+    private final UndeliveredParcelPostService undelivParcelPostService;
+    private final UndeliveredParcelService undeliveredParcelService;
 
     // 게시글 등록
     @PostMapping("/categories/{categoryId}/posts/undelivered-parcel")
@@ -25,7 +27,7 @@ public class UnDeliveredParcelController {
     ) {
         // 이미지 저장
         String email = PrincipalUtil.getEmail(principal);
-        long unDeliParcelId = unDeliParcelService.create(categoryId, email);
+        long unDeliParcelId = undelivParcelPostService.create(categoryId, email);
         ResTemplate<Void> resTemplate = new ResTemplate<>(HttpStatus.CREATED, unDeliParcelId + "번 게시글 작성");
 
         return ResponseEntity
@@ -33,5 +35,18 @@ public class UnDeliveredParcelController {
                 .body(resTemplate);
     }
 
-    // 게시글 내부 관리목록에 관리사항 등록
+//    // 게시글 내부 관리목록에 관리물품 등록
+//    @PostMapping("/categories/{categoryId}/posts/undelivered-parcel/{undelivParcelPostId}/register")
+//    public ResponseEntity<ResTemplate<Void>> handleCreateRegister(
+//            @PathVariable Long categoryId, @PathVariable Long undelivParcelPostId
+//            , @RequestBody
+//    ) {
+//        undeliveredParcelService.create(undelivParcelPostId);
+//        ResTemplate<Void> resTemplate = new ResTemplate<>(HttpStatus.CREATED, undelivParcelPostId + "번 게시글 작성");
+//
+//        return ResponseEntity
+//                .created(URI.create("/categories/"+categoryId+"/posts/undelivered-parcel/" + undelivParcelPostId))
+//                .body(resTemplate);
+//    }
+
 }
