+++ src/main/java/com/dominest/dominestbackend/domain/post/undeliveredparcel/UndeliveredParcelPostRepository.java
@@ -1,0 +1,6 @@
+package com.dominest.dominestbackend.domain.post.undeliveredparcel;
+
+import org.springframework.data.jpa.repository.JpaRepository;
+
+public interface UndeliveredParcelPostRepository extends JpaRepository<UndeliveredParcelPost, Long> {
+}
