--- src/main/java/com/dominest/dominestbackend/domain/post/undeliveredparcelregister/UnDeliveredParcelRegisterPostRepository.java
@@ -1,6 +1,0 @@
-package com.dominest.dominestbackend.domain.post.undeliveredparcelregister;
-
-import org.springframework.data.jpa.repository.JpaRepository;
-
-public interface UnDeliveredParcelRegisterPostRepository extends JpaRepository<UnDeliveredParcelRegisterPost, Long> {
-}
